pyproject.toml
[project]
name = "srt-gen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ffmpeg-python>=0.2.0",
    "openai-whisper>=20231117",
    "ruff>=0.15.7",
    "streamlit>=1.55.0",
    "watchdog>=6.0.0",
]
